Inline Experience component into Actor
