+++ pyro/components/__init__.py
@@ -1,3 +1,2 @@
-from experience import Experience
 from item import Item, Inventory, Equipment, ItemUse, SpellItemUse
 from ai import AI
