+++ pyro/objects.py
@@ -1,7 +1,7 @@
 import json
 import tcod as libtcod
 import pyro.components.ai
-from pyro.components import Experience, Item, Equipment, Inventory
+from pyro.components import Item, Equipment, Inventory
 from pyro.components import SpellItemUse
 from pyro.engine.glyph import Glyph
 from pyro.spells import Confuse, Fireball, Heal, LightningBolt
@@ -35,17 +35,17 @@ def instantiate_spell(template):
 
 def instantiate_monster(template, game):
     name = template['name']
-    exp_comp = Experience(template['experience'])
     spells = None
     if 'spell' in template:
         spells = [instantiate_spell(template['spell'])]
     elif 'spells' in template:
         spells = [instantiate_spell(spell) for spell in template['spells']]
     ai_comp = pyro.components.ai.new(template['ai'], spells)
-    components = [ai_comp, exp_comp]
+    components = [ai_comp]
     game_object = GameObject(name=name, components=components, game=game)
     monster = Monster(game, game_object)
     monster.glyph = Glyph(template['glyph'], getattr(libtcod, template['color']))
+    monster.xp = template['experience']
     monster.hp = template['hp']
     monster.base_max_hp = monster.hp
     monster.base_defense = template['defense']
@@ -89,7 +89,6 @@ def instantiate_item(template):
 
 def make_player(game):
     components = [
-        Experience(xp=0, level=1),
         Inventory(items=[])
     ]
     player = GameObject('Player', components, game=game)
